+++ amt_augmentor/__init__.py
@@ -1,2 +1,2 @@
 # amt_augmentor/__init__.py
-__version__ = "1.2.0"
+__version__ = "1.2.1"